chore: Export the configured `<Navigation />` from `navigation/index`
